+++ cpp-Raytracer/Material.h
@@ -11,10 +11,10 @@
 class Material {
 
 public:
-    float ambient = 0.1;
-    float diffuse = 0.9;
-    float specular = 0.9;
-    float shininess = 200.0;
+    double ambient = 0.1;
+    double diffuse = 0.9;
+    double specular = 0.9;
+    double shininess = 200.0;
     Color color{1, 1, 1};
 
     bool operator==(const Material& other) const;
